Extract enumeration to avoid ts-jest error
