+++ electron-app/src/components/accounts/AccountForm.vue
@@ -58,7 +58,7 @@ import { Component, Emit, Prop, Vue, Watch } from 'vue-property-decorator';
 import { createNamespacedHelpers } from 'vuex';
 import TagInput from '@/components/inputs/TagInput.vue';
 import TagManager from '@/components/tags/TagManager.vue';
-import { TaskType } from '@/model/task';
+import { TaskType } from '@/model/task-type';
 import { BlockchainAccountPayload } from '@/store/balances/actions';
 import { notify } from '@/store/notifications/utils';
 import {
