+++ electron-app/src/model/task.ts
@@ -1,3 +1,4 @@
+import { TaskType } from '@/model/task-type';
 import { taskManager } from '@/services/task-manager';
 import { Blockchain } from '@/typing/types';
 
@@ -29,19 +30,6 @@ export const createTask: <T extends TaskMeta>(
   type,
   meta
 });
-
-export enum TaskType {
-  ADD_ACCOUNT = 'add_account',
-  REMOVE_ACCOUNT = 'remove_account',
-  TRADE_HISTORY = 'process_trade_history',
-  QUERY_BLOCKCHAIN_BALANCES = 'query_blockchain_balances_async',
-  QUERY_EXCHANGE_BALANCES = 'query_exchange_balances_async',
-  QUERY_BALANCES = 'query_balances_async',
-  DSR_BALANCE = 'dsr_balance',
-  DSR_HISTORY = 'dsr_history',
-  MAKEDAO_VAULTS = 'makerdao_vaults',
-  MAKERDAO_VAULT_DETAILS = 'makerdao_vault_details'
-}
 
 export function taskCompletion<R, M extends TaskMeta>(
   task: TaskType
