+++ electron-app/src/store/tasks/getters.ts
@@ -1,7 +1,8 @@
 import find from 'lodash/find';
 import toArray from 'lodash/toArray';
 import { GetterTree } from 'vuex';
-import { Task, TaskMeta, TaskType } from '@/model/task';
+import { Task, TaskMeta } from '@/model/task';
+import { TaskType } from '@/model/task-type';
 import { RotkehlchenState } from '@/store/store';
 import { TaskState } from '@/store/tasks/state';
